Create usePosts custom hook with provider values logic
