+++ src/post/context/PostsContextProvider.tsx
@@ -1,25 +1,10 @@
-import { PropsWithChildren, useCallback, useMemo, useState } from "react";
-import { Post } from "../types";
+import { PropsWithChildren } from "react";
 import { PostsContext } from "./PostsContext";
-import PostClient from "../client/PostClient";
 import PostsContextStructure from "./types";
-import { PostsData } from "../client/types";
+import usePosts from "../hooks/usePosts";
 
 const PostsContextProvider: React.FC<PropsWithChildren> = ({ children }) => {
-  const [posts, setPosts] = useState<Post[]>([]);
-  const [postsTotal, setPostsTotal] = useState<number>(0);
-
-  const postClient = useMemo(() => new PostClient(), []);
-
-  const loadPostsByPage = useCallback(
-    async (pageNumber?: number): Promise<void> => {
-      const postsData: PostsData = await postClient.getPosts(pageNumber);
-
-      setPosts(postsData.posts);
-      setPostsTotal(postsData.postsTotal);
-    },
-    [postClient],
-  );
+  const { loadPostsByPage, posts, postsTotal } = usePosts();
 
   const postsContextValue: PostsContextStructure = {
     posts,
